perf(viewer): cache smoke/fire effects as offscreen sprites
Smoke and fire were the most expensive things the 2D map painted: blur
filters, many radial gradients and additive blending, all rebuilt from
scratch every frame. On weak machines that blew the per-frame budget the
moment a grenade went off, stuttering the replay.

Each effect is now baked once into an offscreen canvas and just blitted
(drawImage) every frame, regenerating only ~12x/sec to keep the subtle
wobble/flicker. The blit scales the sprite to the current zoom radius, so
it stays correct without re-rasterizing the gradients.

While here, split the now-1k-too-long ViewerMap.vue into cohesive modules
(verbatim moves, no behavior change):
- grenadeEffects.ts: the sprite cache + smoke/fire/scorch painters
- canvasUtils.ts: clamp, drawKill, roundRect, wrapText
- useViewerAssets.ts: preloaded mic/weapon/comment-kind icons

ViewerMap.vue: 2687 -> 2406 lines.
